removed "circular includes" by refactoring into cpp and hpp files
