+++ abstract_syntax_tree/ASTRootNode.hpp
@@ -1,51 +1,27 @@
 #ifndef H2_ASTROOTNODE_HPP
 #define H2_ASTROOTNODE_HPP
 
-#include "ASTNode.hpp"
 #include <vector>
+#include <string>
+
+#include "ASTNode.hpp"
+//#include "ASTFunctionNode.hpp"
+struct ASTFunctionNode;
 
 struct ASTRootNode : ASTNode {
-
     std::vector<ASTFunctionNode*> funcs;
 
-    ASTRootNode(std::vector<ASTFunctionNode*>& funcs): funcs(funcs) {}
+    ASTRootNode(std::vector<ASTFunctionNode*>& funcs);
 
-    std::string compile() {
-        std::string code = "jmp main\n\n"; // always starting with main function
+    std::string compile() override;
 
-        for (ASTNode* f : funcs) {
-            code += f->compile();
-        }
+    VirtualMathTerm calculate_complexity() override;
 
-        return code;
-    }
+    std::string to_code() override;
 
-    VirtualMathTerm calculate_complexity() override{
+    void set_block_level(int n) override;
 
-        // sub calls
-        for (auto e : funcs) {
-            e->calculate_complexity();
-        }
-
-        // only functions eg have meaningful complexity. Not the root node itself
-        complexity = VirtualMathTerm("-"); // placeholder
-        return complexity;
-    }
-
-    std::string to_code() override {
-        std::string res;
-        for (auto e : funcs) {
-            res += e->to_code() + "\n\n";
-        }
-        return res;
-    }
-
-    void set_block_level(int n) {
-        block_level = 0;
-        for (auto f: funcs) f->set_block_level(0);
-    }
-
-    std::string get_class() override { return "Root";}
+    std::string get_class() override;
 };
 
 #endif //H2_ASTROOTNODE_HPP
