--- code_style.cpp
@@ -1,4 +1,0 @@
-//
-// Created by ray on 19.01.23.
-//
-#include "code_style.hpp"
