--- code_style.hpp
@@ -1,14 +1,0 @@
-//
-// Created by ray on 13.01.23.
-//
-
-#ifndef H2_CODE_STYLE_HPP
-#define H2_CODE_STYLE_HPP
-
-#include <string>
-
-int block_indention = 4;
-
-
-
-#endif //H2_CODE_STYLE_HPP
