--- complexity_term.cpp
@@ -1,15 +1,0 @@
-//
-// Created by ray on 12.01.23.
-//
-
-#include "complexity_term.hpp"
-
-//VirtualMathTerm::VirtualMathTerm(ComplexityTermType type, std::string name): type(type), name(name) {}
-
-
-// TODO implement some simplification rules (later)
-/*void VirtualMathTerm::simplify() {
-    if (type == ADDITION) {
-
-    }
-}*/
